Fix a unit test related to extensions-of-extensions
The comment says that it segfaulted before. I can't reproduce that
on master and my guess it that it only happened on the release-0.3
branch where I cherry-picked the commit introducing that comment.

This just fixes an issue with the computation of length(...) which
explained one of the tests failing.