Remove a wrong TODO
I think I wrote this assuming `pos` in `posmod(x, y)` refers to
x, but it actually refers to `y`. So I don't need to worry
about this TODO.

It's still possible to optimize this in case typemin(x) >= 0
and also typemax(x) <= char(F), but I don't think that actually
happens a lot in practice.