Add code for computing n-th roots of unity
In many application of finite fields of practical interest, a common
need to compute the an n-th root of unity in this finite field. This
provides (randomized) algorithms for computing either an arbitrary
n-th root or the unique minimum n-th root. Computing the latter is
expensive and of little mathematical significance, but due to the
randomized nature of the algorithm, users may want predictability
in which root gets chosen.