Allow for larger integers when BitIntegers is also loaded
This is a slight addition to @keno's work. He noticed that most
code continuous to work even for bigger integers but was hesitant
to add a new dependency. So am I, but the Requires.jl package caters
to exactly that situation: we extend the allowed primes when
the user chooses to also load BitIntegers.